requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.1.7
django-cors-headers==3.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
